SPDM Responder daemon
This implements SPDM Responder daemon that responds to
SPDM requester messages that are send to attest BMC.
It integrates with spdmapplib to receive and send SPDM messages,
also to format the responses. The daemon starts by reading the
spdm configuration from entity manager and initializing
SPDM responder configuration.

Tested:
Tested with aardvark scripts acting as SPDM requester.
BMC was able to respond to all the SPDM request messages.

Co-authored-by: hsiaoh1x <hsiaoh1x@intel.com>
Co-authored-by: yihuangx <yi-huangx.han@intel.com>

Signed-off-by: Vikram Bodireddy <vikram.bodireddy@intel.com>